makes BNRRandomBytes' length arg a size_t

+++ BNRPersistence/BNRCrypto.h
@@ -6,4 +6,5 @@
 //  Copyright 2010 Big Nerd Ranch. All rights reserved.
 //
 
-void BNRRandomBytes(void *buffer, int length);
+/*! Fills |buffer| with |length| random bytes. */
+void BNRRandomBytes(void *buffer, size_t length);
